Added Performance Test 4
This code was used to read the fuel light and push the corresponding
button. A method was added to determine if the fuel light was red or
blue. To save lines, new drive methods with ‘_sleep’ at their ends were
created. The ‘_sleep’ methods are identical to their parent drive
methods except that they add a sleep statement of 300 ms at the end.