Added code used for the individual competition
This code was used on April 1st 2016 for the individual portion of the
robot competition. The code here includes the new and improved drive
methods and features greater robot precision and faster performance.
The main file has been reduced and simplified. Lots of new code has
been added, and much of it requires better organization, so look out
for cleaner code versions for the final competition.